+++ tests/test_api_search_reference.py
@@ -411,7 +411,7 @@ class TestApiSearchReference(object):  # pylint: disable=too-many-public-methods
         assert result.headers == expect_headers
         Content.assert_restapi(result.json, expect_body)
 
-    @pytest.mark.parametrize('server', [['--server', '-q']], indirect=True)
+    @pytest.mark.parametrize('server', [['--server-host', 'localhost:8080', '-q']], indirect=True)
     @pytest.mark.usefixtures('default-references')
     def test_api_search_reference_012(self, server):
         """Search reference with GET.