+++ tests/test_api_search_solution.py
@@ -555,7 +555,7 @@ class TestApiSearchSolution(object):  # pylint: disable=too-many-public-methods
         assert result.headers == expect_headers
         Content.assert_restapi(result.json, expect_body)
 
-    @pytest.mark.parametrize('server', [['--server', '-q']], indirect=True)
+    @pytest.mark.parametrize('server', [['--server-host', 'localhost:8080', '-q']], indirect=True)
     @pytest.mark.usefixtures('default-solutions')
     def test_api_search_solution_016(self, server):
         """Search solution with GET.